Reworked the `CreatureEatAnimation` to work for every of the action animations. The animation it plays is now configured using an enum.
